sysdig: reorganize files to make test parsers happy
